Add support for IConfiguration when using netstandard.
